Force set /etc/hosts in bonnyci images
I'm getting problems with the hostname not being resolvable in sudo.
Normally this just gives a warning message but recently it's been
causing a ~60 second delay to sudo whilst it tries to ping that
hostname.

The hostname should be resolvable anyway so lets fix that in our image.
This may not be applicable to future non-ubuntu images.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>