Link resolv.conf to generated resolvconf
The resolv.conf in place at the end of dib is one copied across from the
builder. This works in situations where the builder and the node are run
in similar environments, but if they are different can lead to dns
timeouts and ansible failures when sudo tries to resolve the hostname.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>