Revert "Link resolv.conf to generated resolvconf"
This reverts commit 14d96d90d7f736d5a6bc9e287a614d4fa8437503.

dib handles resolv.conf specially so it can install things whilst on the
image. The way -infra handles this is to install unbound and link
resolv.conf via rc.local. This is painful.

I really don't want to do it this way, but lets see if we can do it via
start a different way.